requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
certifi==2024.6.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cryptography==42.0.7; python_version >= '3.7'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==5.0.6; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-debug-toolbar==4.4.2; python_version >= '3.8'
django-filter==24.2; python_version >= '3.8'
django-templated-mail==1.1.1
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-csv==3.0.2
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djangorestframework-xml==2.0.0; python_version >= '3.5'
djangorestframework-yaml==2.0.0; python_version >= '3.5'
djoser==2.2.2; python_version >= '3.8' and python_version < '4.0'
idna==3.7; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
pycparser==2.22; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pymysql==1.1.1; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
python3-openid==3.2.0
pyyaml==6.0.1; python_version >= '3.6'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
social-auth-app-django==5.4.1; python_version >= '3.8'
social-auth-core==4.5.4; python_version >= '3.8'
sqlparse==0.5.0; python_version >= '3.8'
tzdata==2024.1; sys_platform == 'win32'
urllib3==2.2.1; python_version >= '3.8'
